requirements.txt
aws-cdk.aws_appsync==1.95.0
aws-cdk.aws_dynamodb==1.95.0
aws_cdk.aws_lambda_python==1.95.0
aws_cdk.aws_dynamodb==1.95.0